requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-RESTful
Flask-Bcrypt
pydantic
PyJWT
pymongo
rich
gunicorn
boto3
Flask-APScheduler
flask_mail
sentry-sdk
python-dotenv
arrow